refactor: 커스텀 예외 구조 수정 및 테스트코드 패키징 잘못된 부분 개선
