--- src/main/java/request/DataIsNotFormDataException.java
@@ -1,8 +1,0 @@
-package request;
-
-class DataIsNotFormDataException extends RuntimeException {
-
-    DataIsNotFormDataException(String message) {
-        super(message);
-    }
-}
