+++ src/main/java/request/MessageBody.java
@@ -14,12 +14,12 @@ class MessageBody {
 
     String findDataFromFormFormatBody(String fieldName) {
         if (Objects.isNull(messageBody) || messageBody.isEmpty()) {
-            throw new DataIsNotFormDataException("message body is empty.");
+            throw new RequestDataFormatException("message body is empty.");
         }
         try {
             return new QueryData(messageBody).getValue(fieldName);
         } catch (IllegalArgumentException e) {
-            throw new DataIsNotFormDataException("message body is not form data format.");
+            throw new RequestDataFormatException("message body is not form data format.");
         }
     }
