+++ src/main/java/request/RequestDataFormatException.java
@@ -1,0 +1,8 @@
+package request;
+
+public class RequestDataFormatException extends RuntimeException {
+
+    RequestDataFormatException(String message) {
+        super(message);
+    }
+}
