--- src/main/java/request/WrongRequestDataFormatException.java
@@ -1,8 +1,0 @@
-package request;
-
-public class WrongRequestDataFormatException extends RuntimeException {
-
-    WrongRequestDataFormatException(String message) {
-        super(message);
-    }
-}
