+++ src/test/java/request/QueryDataTest.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -10,7 +10,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import request.QueryData;
 
 class QueryDataTest {
