+++ src/test/java/resource/ResourcesHandlerTest.java
@@ -1,4 +1,4 @@
-package webserver.resource;
+package resource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -7,9 +7,6 @@ import java.io.IOException;
 import java.net.URISyntaxException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import resource.ContentType;
-import resource.Resource;
-import resource.ResourcesHandler;
 
 class ResourcesHandlerTest {
